pyproject.toml
[project]
name = "data-maprva-org"
version = "0.1.0"
description = "Data portal for Richmond, VA and the surrounding area."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "geopandas>=1.1.1",
    "osm2geojson>=0.2.6",
    "pyarrow>=21.0.0",
    "pyyaml>=6.0.2",
    "shapely>=2.1.1",
]
